alg/scalable-birl/requirements.txt
dm-haiku @ git+https://github.com/deepmind/dm-haiku@7e84ed108fcc644f3a65327fba818bf20b225a8e
gym==0.17.2
jax==0.1.72
jaxlib==0.1.48
numpy==1.19.1
tqdm==4.48.0